2015/06 in Go

+++ go/2015/02/solution.go
@@ -9,7 +9,7 @@ import (
 func main() {
 	lines := utils.ReadLines()
 
-	var areaAcc, ribbonAcc int64
+	var areaAcc, ribbonAcc int
 	areaAcc = 0
 	ribbonAcc = 0
